refactor(tests): relocate #[cfg(test)] blocks next to their extracted modules (#224)
Move the unit tests for the #222 / #225-extracted modules out of lib.rs's giant
`mod tests` into a `#[cfg(test)] mod tests` inside the module that owns the code
under test. Pure relocation, no assertion changes. lib.rs drops ~930 lines
(now ~2.66k).

- url.rs          <- github_owner_repo / installed_full_name / urls_match / format_plugin_url
- merge.rs        <- ensure_absent / decide_merge_mode (folded into its existing mod tests)
- paths.rs        <- expand_tilde / is_valid_appname / resolve_* / change_record_from / record_changes_or_warn
- plugin_build.rs <- collect_denops / parse_build_command / collect_build_rtp / unwrap_anonymous_lua_function / run_build_lua (+ the shared `write_file` test helper)

Each relocated block uses `use super::*` plus the minimal extra imports it needs.
Dropped the now-unused `use crate::merge::*` / `use crate::plugin_build::*` globs
from lib.rs's mod tests. Tests for code that still lives in lib.rs stay put.

Behaviour-preserving: 822 tests pass (unchanged), plural doctest preserved.
cargo make check green: clippy -D warnings --all-targets, fmt, lock-check.

Out of scope (small follow-up): the completion-generation test (now testing
cli.rs) and the parse_config tests (config.rs) — not #222/#225 leftovers.

Refs #224, #217.
