+++ src/lib.rs
@@ -1206,66 +1206,17 @@ impl Drop for LoaderSwapGuard {
 #[cfg(test)]
 mod tests {
     use super::*;
-    // These helpers moved to dedicated modules (#217); the test cases here still
-    // reference them by bare name, so glob them in until the test blocks are
-    // relocated next to their modules (tracked in #224).
+    // Tests for code that still lives in `lib.rs`. Tests for the extracted
+    // `paths` / `merge` / `url` / `plugin_build` modules were relocated next to
+    // those modules in #224; the `url` glob stays for the few lib.rs tests that
+    // still call into it.
     use crate::config::{Config, MapSpec, Options, Plugin};
     use crate::loader::PluginScripts;
-    use crate::merge::*;
-    use crate::plugin_build::*;
     use crate::url::*;
     use tempfile::tempdir;
     use toml_edit::DocumentMut;
 
     // ── ensure_absent: 残骸混入を防ぐ事前 cleanup ────────────────────
-
-    #[test]
-    fn ensure_absent_removes_existing_dir() {
-        let tmp = tempdir().unwrap();
-        let dir = tmp.path().join("stale");
-        std::fs::create_dir_all(dir.join("sub")).unwrap();
-        std::fs::write(dir.join("sub/file.txt"), b"residue").unwrap();
-        ensure_absent(&dir).unwrap();
-        assert!(!dir.exists(), "dir should be gone after ensure_absent");
-    }
-
-    #[test]
-    fn ensure_absent_is_ok_when_missing() {
-        let tmp = tempdir().unwrap();
-        let dir = tmp.path().join("never-existed");
-        // 存在しない path でも Ok を返す (べき等)。
-        ensure_absent(&dir).unwrap();
-        assert!(!dir.exists());
-    }
-
-    #[test]
-    fn ensure_absent_removes_plain_file() {
-        // 残骸がディレクトリではなくファイルのケース (#223)。Windows で
-        // remove_dir_all を file に対して呼ぶと失敗するので型分岐が要る。
-        let tmp = tempdir().unwrap();
-        let file = tmp.path().join("stale-file");
-        std::fs::write(&file, b"residue").unwrap();
-        ensure_absent(&file).unwrap();
-        assert!(!file.exists(), "file should be gone after ensure_absent");
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn ensure_absent_removes_dangling_symlink() {
-        // dangling symlink: target は存在しないが symlink 自体は残骸として残る (#223)。
-        let tmp = tempdir().unwrap();
-        let link = tmp.path().join("dangling");
-        std::os::unix::fs::symlink(tmp.path().join("missing-target"), &link).unwrap();
-        // symlink 自体は存在する (target は無い)。
-        assert!(link.symlink_metadata().is_ok());
-        ensure_absent(&link).unwrap();
-        assert!(
-            link.symlink_metadata().is_err(),
-            "dangling symlink should be gone after ensure_absent"
-        );
-    }
-
-    // ── patch_plugin_entry_triggers regression: respect existing user fields ──
 
     #[test]
     fn patch_plugin_entry_triggers_does_not_overwrite_existing_on_cmd() {
@@ -1682,13 +1633,6 @@ url = "only/one"
         assert!(!msg.contains("snacks"));
     }
 
-    fn write_file(path: &Path, content: &str) {
-        if let Some(parent) = path.parent() {
-            std::fs::create_dir_all(parent).unwrap();
-        }
-        std::fs::write(path, content).unwrap();
-    }
-
     // ─── classify_held_back ──────────────────────────────────────────────
     // Pure classification of "this plugin is being held back by a lockfile
     // pin". The integration path is covered by the git::remote_head test
@@ -1747,113 +1691,6 @@ url = "only/one"
         // Same resilience argument in the other direction.
         let got = classify_held_back(None, Some("abc"), None, Some("def"));
         assert_eq!(got, None);
-    }
-
-    #[test]
-    fn test_collect_denops_plugins_finds_main_ts() {
-        let root = tempdir().unwrap();
-        let plugin = root.path().join("plugin-repo");
-        write_file(
-            &plugin.join("denops/foo/main.ts"),
-            "export async function main() {}",
-        );
-        write_file(&plugin.join("denops/foo/util.ts"), "export const x = 1;");
-
-        let got = collect_denops_plugins(&plugin);
-        assert_eq!(got.len(), 1);
-        assert_eq!(got[0].name, "foo");
-        assert!(
-            got[0].main_script.ends_with("denops/foo/main.ts"),
-            "main_script should be absolute path ending with denops/foo/main.ts, got: {}",
-            got[0].main_script
-        );
-        // forward slash に正規化されている
-        assert!(
-            !got[0].main_script.contains('\\'),
-            "main_script must use forward slashes"
-        );
-    }
-
-    #[test]
-    fn test_collect_denops_plugins_returns_empty_without_denops_dir() {
-        let root = tempdir().unwrap();
-        let plugin = root.path().join("plugin-repo");
-        write_file(&plugin.join("plugin/foo.vim"), "echo 'foo'");
-        // denops/ が無い
-        let got = collect_denops_plugins(&plugin);
-        assert!(got.is_empty());
-    }
-
-    #[test]
-    fn test_collect_denops_plugins_falls_back_to_main_js() {
-        let root = tempdir().unwrap();
-        let plugin = root.path().join("plugin-repo");
-        // main.ts なし、main.js のみ
-        write_file(
-            &plugin.join("denops/bar/main.js"),
-            "export async function main() {}",
-        );
-
-        let got = collect_denops_plugins(&plugin);
-        assert_eq!(got.len(), 1);
-        assert_eq!(got[0].name, "bar");
-        assert!(got[0].main_script.ends_with("denops/bar/main.js"));
-    }
-
-    #[test]
-    fn test_collect_denops_plugins_prefers_main_ts_over_main_js() {
-        let root = tempdir().unwrap();
-        let plugin = root.path().join("plugin-repo");
-        write_file(&plugin.join("denops/dual/main.ts"), "ts");
-        write_file(&plugin.join("denops/dual/main.js"), "js");
-        let got = collect_denops_plugins(&plugin);
-        assert_eq!(got.len(), 1);
-        assert!(got[0].main_script.ends_with("main.ts"));
-    }
-
-    #[test]
-    fn test_collect_denops_plugins_skips_dirs_without_main() {
-        let root = tempdir().unwrap();
-        let plugin = root.path().join("plugin-repo");
-        // main.ts も main.js も無いディレクトリは無視
-        write_file(&plugin.join("denops/incomplete/other.ts"), "");
-        write_file(&plugin.join("denops/ok/main.ts"), "");
-        let got = collect_denops_plugins(&plugin);
-        assert_eq!(got.len(), 1);
-        assert_eq!(got[0].name, "ok");
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn test_collect_denops_plugins_follows_symlinked_subdir() {
-        // dev plugin や mono-repo の構成で、denops/<name>/ が別ディレクトリへの
-        // symlink になっているケースを follow して検出する。
-        // Windows は symlink 作成に管理者権限が要るので Unix のみで実行。
-        let root = tempdir().unwrap();
-        let plugin = root.path().join("plugin-repo");
-        let real = root.path().join("external").join("real-denops");
-        write_file(&real.join("main.ts"), "export async function main() {}");
-        std::fs::create_dir_all(plugin.join("denops")).unwrap();
-        std::os::unix::fs::symlink(&real, plugin.join("denops/sym-linked")).unwrap();
-
-        let got = collect_denops_plugins(&plugin);
-        assert_eq!(got.len(), 1);
-        assert_eq!(got[0].name, "sym-linked");
-        assert!(got[0].main_script.ends_with("main.ts"));
-    }
-
-    #[test]
-    fn test_collect_denops_plugins_multiple_sorted_by_name() {
-        // 同一プラグイン repo が複数の denops サブモジュールを持つケース
-        // (例: 一部の mono-repo) も決定論的な順序を保証
-        let root = tempdir().unwrap();
-        let plugin = root.path().join("plugin-repo");
-        write_file(&plugin.join("denops/zeta/main.ts"), "");
-        write_file(&plugin.join("denops/alpha/main.ts"), "");
-        write_file(&plugin.join("denops/mid/main.ts"), "");
-        let got = collect_denops_plugins(&plugin);
-        let names: Vec<&str> = got.iter().map(|d| d.name.as_str()).collect();
-        assert_eq!(names, vec!["alpha", "mid", "zeta"]);
     }
 
     #[test]
@@ -1948,215 +1785,6 @@ url = "only/one"
     }
 
     #[test]
-    fn test_installed_full_name_owner_repo() {
-        assert_eq!(
-            installed_full_name("folke/snacks.nvim"),
-            Some("folke/snacks.nvim".to_string())
-        );
-    }
-
-    #[test]
-    fn test_installed_full_name_https_url_with_git_suffix() {
-        assert_eq!(
-            installed_full_name("https://github.com/Owner/Repo.git"),
-            Some("owner/repo".to_string())
-        );
-    }
-
-    #[test]
-    fn test_installed_full_name_https_url_without_git_suffix() {
-        assert_eq!(
-            installed_full_name("https://github.com/nvim-lua/plenary.nvim"),
-            Some("nvim-lua/plenary.nvim".to_string())
-        );
-    }
-
-    #[test]
-    fn test_installed_full_name_ssh_url() {
-        assert_eq!(
-            installed_full_name("git@github.com:Owner/Repo.git"),
-            Some("owner/repo".to_string())
-        );
-    }
-
-    #[test]
-    fn test_installed_full_name_non_github_returns_none() {
-        assert_eq!(installed_full_name("https://gitlab.com/owner/repo"), None);
-    }
-
-    #[test]
-    fn test_installed_full_name_case_normalized() {
-        assert_eq!(
-            installed_full_name("Folke/Snacks.NVIM"),
-            Some("folke/snacks.nvim".to_string())
-        );
-    }
-
-    // -------- urls_match --------
-
-    #[test]
-    fn test_urls_match_treats_github_short_and_full_as_equal() {
-        // `owner/repo` 形式と `https://github.com/owner/repo` 形式は同一リポジトリ扱い。
-        assert!(urls_match(
-            "folke/snacks.nvim",
-            "https://github.com/folke/snacks.nvim",
-        ));
-        assert!(urls_match(
-            "https://github.com/Folke/Snacks.NVIM.git",
-            "folke/snacks.nvim",
-        ));
-        assert!(urls_match(
-            "git@github.com:folke/snacks.nvim.git",
-            "folke/snacks.nvim",
-        ));
-    }
-
-    #[test]
-    fn test_urls_match_rejects_different_owner_or_repo() {
-        // 同じ repo 名でも owner が違うのは別プラグイン。
-        assert!(!urls_match("foo/snacks.nvim", "bar/snacks.nvim"));
-        assert!(!urls_match("folke/snacks.nvim", "folke/other.nvim"));
-    }
-
-    #[test]
-    fn test_urls_match_non_github_falls_back_to_string_eq() {
-        // GitHub と認識できない URL は生文字列比較。
-        assert!(urls_match(
-            "https://gitlab.com/x/y",
-            "https://gitlab.com/x/y",
-        ));
-        assert!(!urls_match(
-            "https://gitlab.com/x/y",
-            "https://gitlab.com/x/z",
-        ));
-    }
-
-    #[test]
-    fn test_installed_full_name_trailing_slash() {
-        // `owner/repo/`, `.../repo.git/`, `.../repo/` をすべて許容する
-        assert_eq!(
-            installed_full_name("folke/snacks.nvim/"),
-            Some("folke/snacks.nvim".to_string())
-        );
-        assert_eq!(
-            installed_full_name("https://github.com/Owner/Repo/"),
-            Some("owner/repo".to_string())
-        );
-        assert_eq!(
-            installed_full_name("https://github.com/Owner/Repo.git/"),
-            Some("owner/repo".to_string())
-        );
-    }
-
-    #[test]
-    fn test_github_owner_repo_preserves_case() {
-        // installed_full_name と違い case は保持する (config.toml 書き込み用途)
-        assert_eq!(
-            github_owner_repo("Folke/Snacks.NVIM"),
-            Some("Folke/Snacks.NVIM".to_string())
-        );
-        assert_eq!(
-            github_owner_repo("https://github.com/Owner/Repo.git"),
-            Some("Owner/Repo".to_string())
-        );
-        assert_eq!(
-            github_owner_repo("git@github.com:Owner/Repo.git"),
-            Some("Owner/Repo".to_string())
-        );
-    }
-
-    #[test]
-    fn test_github_owner_repo_non_github_returns_none() {
-        assert_eq!(github_owner_repo("https://gitlab.com/a/b"), None);
-        assert_eq!(github_owner_repo("https://git.example/a/b"), None);
-    }
-
-    #[test]
-    fn test_format_plugin_url_short_form() {
-        use crate::config::UrlStyle;
-        // 入力がどの形式でも owner/repo に統一
-        assert_eq!(
-            format_plugin_url("folke/snacks.nvim", UrlStyle::Short),
-            "folke/snacks.nvim"
-        );
-        assert_eq!(
-            format_plugin_url("https://github.com/folke/snacks.nvim", UrlStyle::Short),
-            "folke/snacks.nvim"
-        );
-        assert_eq!(
-            format_plugin_url("https://github.com/folke/snacks.nvim.git", UrlStyle::Short),
-            "folke/snacks.nvim"
-        );
-    }
-
-    #[test]
-    fn test_format_plugin_url_full_form() {
-        use crate::config::UrlStyle;
-        // 入力がどの形式でも https://github.com/owner/repo に統一
-        assert_eq!(
-            format_plugin_url("folke/snacks.nvim", UrlStyle::Full),
-            "https://github.com/folke/snacks.nvim"
-        );
-        assert_eq!(
-            format_plugin_url("https://github.com/folke/snacks.nvim.git", UrlStyle::Full),
-            "https://github.com/folke/snacks.nvim"
-        );
-        assert_eq!(
-            format_plugin_url("git@github.com:folke/snacks.nvim.git", UrlStyle::Full),
-            "https://github.com/folke/snacks.nvim"
-        );
-    }
-
-    #[test]
-    fn test_format_plugin_url_non_github_passthrough() {
-        use crate::config::UrlStyle;
-        // GitHub 以外は入力そのまま保存 (style 無視)
-        assert_eq!(
-            format_plugin_url("https://gitlab.com/g/h", UrlStyle::Short),
-            "https://gitlab.com/g/h"
-        );
-        assert_eq!(
-            format_plugin_url("https://gitlab.com/g/h", UrlStyle::Full),
-            "https://gitlab.com/g/h"
-        );
-    }
-
-    #[test]
-    fn test_github_owner_repo_rejects_local_paths() {
-        // `./foo`, `../foo`, `~/foo`, `/foo`, `\foo`, `C:/foo` を GitHub shorthand
-        // と誤認しないこと (CodeRabbit major fix)。
-        assert_eq!(github_owner_repo("./foo"), None);
-        assert_eq!(github_owner_repo("../foo"), None);
-        assert_eq!(github_owner_repo("~/foo"), None);
-        assert_eq!(github_owner_repo("/tmp/foo"), None);
-        assert_eq!(github_owner_repo("\\foo\\bar"), None);
-        assert_eq!(github_owner_repo("C:/foo"), None);
-        assert_eq!(github_owner_repo("d:/bar/baz"), None);
-    }
-
-    #[test]
-    fn test_github_owner_repo_rejects_invalid_chars() {
-        // owner / repo の文字集合を超えるものは GitHub shorthand と認めない。
-        // owner は alphanumeric + `-`、repo は + `. _` まで許容。
-        assert_eq!(github_owner_repo("foo bar/baz"), None);
-        assert_eq!(github_owner_repo("-foo/bar"), None); // owner 先頭 `-`
-        assert_eq!(github_owner_repo("foo!/bar"), None); // owner に `!`
-    }
-
-    #[test]
-    fn test_github_owner_repo_accepts_normal_shorthand() {
-        // 正常な owner/repo は従来どおり受理される
-        assert_eq!(
-            github_owner_repo("folke/snacks.nvim"),
-            Some("folke/snacks.nvim".to_string())
-        );
-        assert_eq!(
-            github_owner_repo("nvim-lua/plenary.nvim"),
-            Some("nvim-lua/plenary.nvim".to_string())
-        );
-    }
-
-    #[test]
     fn test_parse_config_url_style_defaults_to_short() {
         let toml = r#"
 [options]
@@ -2237,137 +1865,6 @@ url = "owner/repo"
             .collect();
         assert_eq!(filtered.len(), 2);
     }
-
-    #[test]
-    fn test_expand_tilde_bare_tilde_returns_home() {
-        let home = dirs::home_dir().unwrap();
-        assert_eq!(expand_tilde("~"), home);
-    }
-
-    #[test]
-    fn test_expand_tilde_with_forward_slash_subpath() {
-        let home = dirs::home_dir().unwrap();
-        assert_eq!(expand_tilde("~/foo/bar"), home.join("foo").join("bar"));
-    }
-
-    #[test]
-    fn test_expand_tilde_with_backslash_subpath() {
-        let home = dirs::home_dir().unwrap();
-        // Windows 入力形式にも対応
-        let got = expand_tilde("~\\foo\\bar");
-        // 実際のパス区切りは OS 依存だが、home 配下に foo と bar を含むかで判定
-        let s = got.to_string_lossy().replace('\\', "/");
-        let expected = home
-            .join("foo")
-            .join("bar")
-            .to_string_lossy()
-            .replace('\\', "/");
-        assert_eq!(s, expected);
-    }
-
-    #[test]
-    fn test_expand_tilde_absolute_path_untouched() {
-        assert_eq!(
-            expand_tilde("/absolute/path"),
-            PathBuf::from("/absolute/path")
-        );
-    }
-
-    #[test]
-    fn test_expand_tilde_relative_path_untouched() {
-        assert_eq!(
-            expand_tilde("relative/path"),
-            PathBuf::from("relative/path")
-        );
-    }
-
-    #[test]
-    fn test_is_valid_appname_rejects_unsafe_values() {
-        assert!(is_valid_appname("nvim"));
-        assert!(is_valid_appname("nvim-test"));
-        assert!(!is_valid_appname(""));
-        assert!(!is_valid_appname("."));
-        assert!(!is_valid_appname(".."));
-        assert!(!is_valid_appname("foo/bar"));
-        assert!(!is_valid_appname("foo\\bar"));
-        assert!(!is_valid_appname("foo\0bar"));
-    }
-
-    #[test]
-    fn test_resolve_cache_root_uses_appname_default() {
-        let home = dirs::home_dir().unwrap();
-        let result = resolve_cache_root(None);
-        // appname は env に依存するので親ディレクトリだけ確認
-        assert!(result.starts_with(home.join(".cache").join("rvpm")));
-    }
-
-    #[test]
-    fn test_resolve_cache_root_expands_tilde() {
-        let home = dirs::home_dir().unwrap();
-        assert_eq!(
-            resolve_cache_root(Some("~/dotfiles/rvpm")),
-            home.join("dotfiles").join("rvpm")
-        );
-    }
-
-    #[test]
-    fn test_resolve_cache_root_accepts_absolute_path() {
-        assert_eq!(
-            resolve_cache_root(Some("/opt/rvpm")),
-            PathBuf::from("/opt/rvpm")
-        );
-    }
-
-    #[test]
-    fn test_resolve_config_root_uses_default_when_none() {
-        let home = dirs::home_dir().unwrap();
-        let result = resolve_config_root(None);
-        // デフォルトは `~/.config/rvpm/<appname>` (plugins は含まない)
-        assert!(result.starts_with(home.join(".config").join("rvpm")));
-        assert!(!result.ends_with("plugins"));
-        // 末尾は appname (env 依存)
-        assert_eq!(
-            result.parent(),
-            Some(home.join(".config").join("rvpm").as_path())
-        );
-    }
-
-    #[test]
-    fn test_resolve_config_root_expands_tilde() {
-        let home = dirs::home_dir().unwrap();
-        assert_eq!(
-            resolve_config_root(Some("~/dotfiles/nvim")),
-            home.join("dotfiles").join("nvim")
-        );
-    }
-
-    #[test]
-    fn test_resolve_config_root_accepts_absolute_path() {
-        assert_eq!(
-            resolve_config_root(Some("/etc/rvpm")),
-            PathBuf::from("/etc/rvpm")
-        );
-    }
-
-    #[test]
-    fn test_resolve_plugin_config_dir_joins_plugins_subdir() {
-        let plugin = config::Plugin {
-            url: "folke/snacks.nvim".to_string(),
-            ..Default::default()
-        };
-        let root = PathBuf::from("/tmp/rvpm");
-        let got = resolve_plugin_config_dir(&root, &plugin);
-        assert_eq!(
-            got,
-            PathBuf::from("/tmp/rvpm")
-                .join("plugins")
-                .join(plugin.canonical_path())
-        );
-    }
-
-    // -----------------------------------------------------------------
-    // rvpm init ヘルパーのテスト
-    // -----------------------------------------------------------------
 
     #[test]
     fn test_nvim_init_lua_path_for_appname_defaults_to_nvim() {
@@ -2536,13 +2033,6 @@ url = "owner/repo"
     }
 
     #[test]
-    fn test_resolve_loader_path_is_under_plugins() {
-        let base = PathBuf::from("/cache/rvpm/nvim");
-        let result = resolve_loader_path(&base);
-        assert_eq!(result, PathBuf::from("/cache/rvpm/nvim/plugins/loader.lua"));
-    }
-
-    #[test]
     fn test_write_loader_to_path_creates_file() {
         let root = tempdir().unwrap();
         let merged = root.path().join("merged");
@@ -2649,81 +2139,6 @@ url = "owner/repo"
                 shell
             );
         }
-    }
-
-    #[test]
-    fn test_decide_merge_mode_eager_merge_is_full() {
-        // eager + merge=true → 常に Full (per-plugin / global の merge_doc は影響しない)
-        assert_eq!(
-            decide_merge_mode(true, false, None, false),
-            PluginMergeMode::Full
-        );
-        assert_eq!(
-            decide_merge_mode(true, false, None, true),
-            PluginMergeMode::Full
-        );
-        assert_eq!(
-            decide_merge_mode(true, false, Some(false), true),
-            PluginMergeMode::Full
-        );
-        assert_eq!(
-            decide_merge_mode(true, false, Some(true), false),
-            PluginMergeMode::Full
-        );
-    }
-
-    #[test]
-    fn test_decide_merge_mode_global_default_lazy() {
-        // lazy + merge=true + per-plugin None → global で決まる
-        assert_eq!(
-            decide_merge_mode(true, true, None, false),
-            PluginMergeMode::ViewWithDoc
-        );
-        assert_eq!(
-            decide_merge_mode(true, true, None, true),
-            PluginMergeMode::ViewWithoutDoc
-        );
-    }
-
-    #[test]
-    fn test_decide_merge_mode_per_plugin_override_wins() {
-        // per-plugin Some(true) は global=false でも勝つ
-        assert_eq!(
-            decide_merge_mode(true, true, Some(true), false),
-            PluginMergeMode::ViewWithoutDoc
-        );
-        // per-plugin Some(false) は global=true でも勝つ
-        assert_eq!(
-            decide_merge_mode(true, true, Some(false), true),
-            PluginMergeMode::ViewWithDoc
-        );
-    }
-
-    #[test]
-    fn test_decide_merge_mode_eager_non_merge_uses_view() {
-        // eager + merge=false: view 経由で扱う (clone path 直 rtp:append しない統一案)
-        assert_eq!(
-            decide_merge_mode(false, false, None, false),
-            PluginMergeMode::ViewWithDoc
-        );
-        // merge_doc=true なら eager + merge=false でも doc を merged/ に集約
-        assert_eq!(
-            decide_merge_mode(false, false, Some(true), false),
-            PluginMergeMode::ViewWithoutDoc
-        );
-    }
-
-    #[test]
-    fn test_decide_merge_mode_lazy_non_merge_uses_view() {
-        // lazy + merge=false: view 経由で扱う
-        assert_eq!(
-            decide_merge_mode(false, true, None, false),
-            PluginMergeMode::ViewWithDoc
-        );
-        assert_eq!(
-            decide_merge_mode(false, true, Some(true), false),
-            PluginMergeMode::ViewWithoutDoc
-        );
     }
 
     #[test]
@@ -3104,257 +2519,6 @@ lazy = false"#;
     // -----------------------------------------------------------------
 
     #[test]
-    fn test_parse_build_command_shell() {
-        let dirs = vec![PathBuf::from("/path/to/plugin")];
-        let (cmd, args) = parse_build_command("cargo build --release", &dirs);
-        if cfg!(windows) {
-            assert_eq!(cmd, "cmd");
-            assert_eq!(args, vec!["/C", "cargo build --release"]);
-        } else {
-            assert_eq!(cmd, "sh");
-            assert_eq!(args, vec!["-c", "cargo build --release"]);
-        }
-    }
-
-    #[test]
-    fn test_parse_build_command_vim_prefix() {
-        let dirs = vec![PathBuf::from("/path/to/plugin")];
-        let (cmd, args) = parse_build_command(":call mkdp#util#install()", &dirs);
-        assert_eq!(cmd, "nvim");
-        assert!(args.iter().any(|a| a == "--headless"));
-        assert!(args.iter().any(|a| a.contains("mkdp#util#install()")));
-    }
-
-    #[test]
-    fn test_parse_build_command_vim_simple() {
-        let dirs = vec![PathBuf::from("/path/to/plugin")];
-        let (cmd, args) = parse_build_command(":TSUpdate", &dirs);
-        assert_eq!(cmd, "nvim");
-        assert!(args.iter().any(|a| a == "--headless"));
-        assert!(args.iter().any(|a| a.contains("TSUpdate")));
-    }
-
-    #[test]
-    fn test_parse_build_command_vim_adds_rtp() {
-        let dirs = vec![PathBuf::from("/path/to/my-plugin")];
-        let (cmd, args) = parse_build_command(":MyBuild", &dirs);
-        assert_eq!(cmd, "nvim");
-        assert!(args.iter().any(|a| a == "--cmd"));
-        assert!(
-            args.iter()
-                .any(|a| a.contains("set rtp+=/path/to/my-plugin")),
-            "should add plugin dir to rtp: {:?}",
-            args
-        );
-    }
-
-    #[test]
-    fn test_parse_build_command_vim_includes_deps_rtp() {
-        let dirs = vec![
-            PathBuf::from("/path/to/plugin"),
-            PathBuf::from("/path/to/dep1"),
-            PathBuf::from("/path/to/dep2"),
-        ];
-        let (cmd, args) = parse_build_command(":Build", &dirs);
-        assert_eq!(cmd, "nvim");
-        let rtp_arg = args
-            .iter()
-            .find(|a| a.contains("set rtp+="))
-            .expect("should have rtp cmd");
-        assert!(rtp_arg.contains("/path/to/plugin"), "self: {}", rtp_arg);
-        assert!(rtp_arg.contains("/path/to/dep1"), "dep1: {}", rtp_arg);
-        assert!(rtp_arg.contains("/path/to/dep2"), "dep2: {}", rtp_arg);
-    }
-
-    // ─── build_lua / collect_build_rtp (#97) ────────────────────────────
-
-    #[test]
-    fn collect_build_rtp_includes_self_first_then_transitive_deps() {
-        // A depends [B], B depends [C]. rtp should be [A, B, C]. (#97)
-        let plugin_a = Plugin {
-            url: "owner/a".to_string(),
-            depends: Some(vec!["b".to_string()]),
-            ..Default::default()
-        };
-        let plugin_b = Plugin {
-            name: Some("b".to_string()),
-            url: "owner/b".to_string(),
-            depends: Some(vec!["c".to_string()]),
-            ..Default::default()
-        };
-        let plugin_c = Plugin {
-            name: Some("c".to_string()),
-            url: "owner/c".to_string(),
-            ..Default::default()
-        };
-        let config = Config {
-            vars: None,
-            options: Options::default(),
-            plugins: vec![plugin_a.clone(), plugin_b, plugin_c],
-        };
-        let cache_root = PathBuf::from("/cache");
-        let rtp = collect_build_rtp(
-            &plugin_a,
-            &PathBuf::from("/cache/plugins/repos/owner/a"),
-            &config,
-            &cache_root,
-        );
-        // self が先頭
-        assert_eq!(rtp[0], PathBuf::from("/cache/plugins/repos/owner/a"));
-        // transitive deps が含まれる (順序は DFS なので厳密には保証しないが、
-        // 全部存在することを確認)
-        let rtp_strings: Vec<String> = rtp.iter().map(|p| p.display().to_string()).collect();
-        assert!(
-            rtp_strings.iter().any(|s| s.contains("owner/b")),
-            "B in rtp: {rtp_strings:?}"
-        );
-        assert!(
-            rtp_strings.iter().any(|s| s.contains("owner/c")),
-            "C in rtp: {rtp_strings:?}"
-        );
-    }
-
-    #[test]
-    fn unwrap_anonymous_lua_function_strips_lazy_nvim_style_wrapper() {
-        // lazy.nvim style — user が `build = function() ... end` から copy
-        let r = unwrap_anonymous_lua_function(
-            "function() require('blink.cmp').build():wait(60000) end",
-        );
-        assert_eq!(
-            r.as_deref(),
-            Some("require('blink.cmp').build():wait(60000)"),
-            "function() body end → body extraction"
-        );
-    }
-
-    #[test]
-    fn unwrap_anonymous_lua_function_handles_space_after_function_keyword() {
-        let r = unwrap_anonymous_lua_function("function ()  vim.cmd('TSUpdate')  end");
-        assert_eq!(r.as_deref(), Some("vim.cmd('TSUpdate')"));
-    }
-
-    #[test]
-    fn unwrap_anonymous_lua_function_returns_none_for_plain_statement() {
-        // 既に statement なら触らない (None で「unwrap 不要」と通知)。
-        let r = unwrap_anonymous_lua_function("require('x').build():wait(60000)");
-        assert!(r.is_none());
-    }
-
-    #[test]
-    fn unwrap_anonymous_lua_function_returns_none_for_named_function_decl() {
-        // `function name() ... end` は valid な statement (function 宣言) なので
-        // 触る理由が無い。
-        let r = unwrap_anonymous_lua_function("function build_step() require('x') end");
-        assert!(r.is_none());
-    }
-
-    #[test]
-    fn unwrap_anonymous_lua_function_does_not_split_identifier_ending_in_end() {
-        // `local end_var = 1 end` のように identifier が "end" で終わるケースで
-        // strip_suffix が誤動作しないこと。strip_suffix は exact suffix match
-        // なので、"end_var" の "end" 部分は拾わない (前後の文字を見るので)。
-        let r = unwrap_anonymous_lua_function("function() local end_var = 1 end");
-        assert_eq!(r.as_deref(), Some("local end_var = 1"));
-    }
-
-    #[test]
-    fn unwrap_anonymous_lua_function_preserves_inner_function_blocks() {
-        // 入れ子の `function() ... end` は inner 側を保持する (lazy `.*?` + 末尾
-        // anchor で「最後の一番外の end」だけを落とす)。
-        let r =
-            unwrap_anonymous_lua_function("function() local f = function() return 1 end; f() end");
-        assert_eq!(r.as_deref(), Some("local f = function() return 1 end; f()"));
-    }
-
-    #[test]
-    fn unwrap_anonymous_lua_function_handles_multiline_body() {
-        // 改行を含む typical な複数行 build_lua。
-        let code = "function()\n  local cmp = require('blink.cmp')\n  cmp.build():wait(60000)\n  vim.print('done')\nend";
-        let r = unwrap_anonymous_lua_function(code);
-        let body = r.expect("multiline function() should unwrap");
-        assert!(body.contains("local cmp = require('blink.cmp')"));
-        assert!(body.contains("cmp.build():wait(60000)"));
-        assert!(body.contains("vim.print('done')"));
-        assert!(!body.contains("function()"));
-        // 末尾に余分な end が残ってないこと
-        assert!(!body.ends_with("end"));
-    }
-
-    #[test]
-    fn unwrap_anonymous_lua_function_handles_complex_multiline_with_nested_function() {
-        // 複雑な複数行: ローカル関数定義入り。inner の function() ... end は保持。
-        let code = "function()\n  local helper = function(x)\n    return x * 2\n  end\n  print(helper(21))\nend";
-        let r = unwrap_anonymous_lua_function(code);
-        let body = r.expect("complex multiline should unwrap outer only");
-        assert!(body.contains("local helper = function(x)"));
-        assert!(body.contains("return x * 2"));
-        assert!(body.contains("print(helper(21))"));
-        // inner の helper 定義 end が残ってる
-        assert!(
-            body.matches("end").count() >= 1,
-            "inner helper end should remain, body: {body}"
-        );
-    }
-
-    #[test]
-    fn unwrap_anonymous_lua_function_tolerates_trailing_line_comment() {
-        // `end` の後に行コメントが付くケース (regex で末尾 `(?:--[^\n]*)?` 許容)。
-        let r = unwrap_anonymous_lua_function("function() vim.print('hi') end -- post-build hook");
-        assert_eq!(r.as_deref(), Some("vim.print('hi')"));
-    }
-
-    #[test]
-    fn unwrap_anonymous_lua_function_tolerates_whitespace_inside_parens() {
-        // `function ( )` のような余分な空白も許容。
-        let r = unwrap_anonymous_lua_function("function(  )  print('x')  end");
-        assert_eq!(r.as_deref(), Some("print('x')"));
-    }
-
-    #[test]
-    fn unwrap_anonymous_lua_function_returns_none_when_function_takes_args() {
-        // `function(a, b)` のように引数があるものは触らない (匿名 zero-arity 専用)。
-        let r = unwrap_anonymous_lua_function("function(a, b) return a + b end");
-        assert!(r.is_none(), "function with args should not be unwrapped");
-    }
-
-    #[tokio::test]
-    async fn run_build_lua_returns_none_when_field_unset() {
-        let plugin = Plugin {
-            url: "x/y".to_string(),
-            ..Default::default()
-        };
-        let result = run_build_lua(&plugin, &PathBuf::from("/tmp"), &[]).await;
-        assert!(
-            result.is_none(),
-            "no build_lua field → no-op, got: {result:?}"
-        );
-    }
-
-    #[tokio::test]
-    async fn run_build_lua_reports_error_when_nvim_missing() {
-        // nvim が PATH に無くても rvpm はクラッシュさせず、明示的なエラー文字列を返す。
-        // この test は nvim がインストールされてる環境ではスキップされる
-        // (nvim コマンドが本当に成功してしまうため)。CI / 開発機の typical 構成では
-        // nvim あり、ローカル run_build_lua の no-build-lua path を検証する別 test
-        // (run_build_lua_returns_none_when_field_unset) で十分。
-        if which::which("nvim").is_ok() {
-            return;
-        }
-        let tmp = tempdir().unwrap();
-        let plugin = Plugin {
-            url: "x/y".to_string(),
-            build_lua: Some("vim.print('hi')".to_string()),
-            ..Default::default()
-        };
-        let result = run_build_lua(&plugin, tmp.path(), &[tmp.path().to_path_buf()]).await;
-        let err = result.expect("missing nvim should yield an error");
-        assert!(
-            err.contains("failed to spawn") || err.contains("nvim"),
-            "error should mention nvim: {err}"
-        );
-    }
-
-    #[test]
     fn test_find_unused_repos() {
         let root = tempdir().unwrap();
         // `cache_root` を root にして、標準の {cache_root}/plugins/repos/ 下に配置する
@@ -3488,97 +2652,4 @@ url = "owner/repo"
     // ================================================================
     // update_log wiring
     // ================================================================
-
-    #[test]
-    fn test_resolve_update_log_path_is_cache_root_sibling() {
-        let cache = std::path::PathBuf::from("/tmp/cache_root");
-        let p = resolve_update_log_path(&cache);
-        assert_eq!(
-            p,
-            std::path::PathBuf::from("/tmp/cache_root/update_log.json")
-        );
-        // loader.lua の親 (= /tmp/cache_root/plugins/) には居ない
-        let loader = resolve_loader_path(&cache);
-        assert_ne!(p.parent(), loader.parent());
-    }
-
-    #[test]
-    fn test_change_record_from_preserves_fields() {
-        let plugin = Plugin {
-            url: "folke/snacks.nvim".to_string(),
-            name: Some("snacks".to_string()),
-            ..Default::default()
-        };
-        let change = crate::git::GitChange {
-            from: Some("aaaaaaa".to_string()),
-            to: "bbbbbbb".to_string(),
-            subjects: vec!["fix: x".to_string()],
-            breaking_subjects: vec![],
-            doc_files_changed: vec!["README.md".to_string()],
-        };
-        let rec = change_record_from(&plugin, change);
-        assert_eq!(rec.name, "snacks");
-        assert_eq!(rec.url, "folke/snacks.nvim");
-        assert_eq!(rec.from.as_deref(), Some("aaaaaaa"));
-        assert_eq!(rec.to, "bbbbbbb");
-        assert_eq!(rec.subjects, vec!["fix: x".to_string()]);
-        assert_eq!(rec.doc_files_changed, vec!["README.md".to_string()]);
-    }
-
-    #[test]
-    fn test_change_record_from_uses_default_name_when_no_name() {
-        let plugin = Plugin {
-            url: "https://github.com/owner/Repo.git".to_string(),
-            ..Default::default()
-        };
-        let change = crate::git::GitChange {
-            from: None,
-            to: "deadbee".to_string(),
-            subjects: vec![],
-            breaking_subjects: vec![],
-            doc_files_changed: vec![],
-        };
-        let rec = change_record_from(&plugin, change);
-        // default_name は `.git` を剥がして最終セグメントを返す
-        assert_eq!(rec.name, "Repo");
-        assert!(rec.from.is_none());
-    }
-
-    #[test]
-    fn test_record_changes_or_warn_writes_file() {
-        // 空 changes は file 自体を作らない方針 (有用履歴を押し出さないため)。
-        // 非空 changes 1 件渡したらきちんと書かれることを確認。
-        let dir = tempdir().unwrap();
-        let cache_root = dir.path().to_path_buf();
-        record_changes_or_warn(
-            &cache_root,
-            "sync",
-            vec![crate::update_log::ChangeRecord {
-                name: "x".into(),
-                url: "owner/x".into(),
-                from: Some("a".into()),
-                to: "b".into(),
-                subjects: vec!["fix: x".into()],
-                breaking_subjects: vec![],
-                doc_files_changed: vec![],
-            }],
-        );
-        let path = resolve_update_log_path(&cache_root);
-        assert!(path.exists(), "expected log file for non-empty changes");
-        let log = crate::update_log::load_log(&path);
-        assert_eq!(log.runs.len(), 1);
-        assert_eq!(log.runs[0].command, "sync");
-    }
-
-    #[test]
-    fn test_record_changes_or_warn_skips_empty() {
-        let dir = tempdir().unwrap();
-        let cache_root = dir.path().to_path_buf();
-        record_changes_or_warn(&cache_root, "sync", vec![]);
-        let path = resolve_update_log_path(&cache_root);
-        assert!(
-            !path.exists(),
-            "empty changes should not create the log file"
-        );
-    }
 }
